refactor(board): 인프라 모듈 분리
