+++ board/board-web-mvc/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    id("seatliberator.web-mvc-base")
+}
+
+group = "com.seatliberator.seatliberator"
+version = "0.0.1-SNAPSHOT"
+
+dependencies {
+    implementation(project(":board:board-persistence"))
+    implementation(project(":board:board-application"))
+    implementation(project(":board:board-domain"))
+
+    implementation(project(":bootstrap:resource-application-starter"))
+
+    testImplementation(project(":kernel:kernel-test"))
+}
