+++ board/board-web-mvc/src/main/java/com/seatliberator/seatliberator/board/BoardApplication.java
@@ -1,0 +1,11 @@
+package com.seatliberator.seatliberator.board;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BoardApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(BoardApplication.class, args);
+    }
+}
