+++ board/board-web-mvc/src/test/resources/application-test.yaml
